Move UserViewSet and UserSerializer into authentication app
urls.py 파일에 들어 있는 것을 제자리 찾아줬음.
router는 지금 남아 있는데,
url root가 users를 출력하도록 되어 있어서 일단 건드리지 않음.
