+++ apps/authentication/serializers.py
@@ -1,0 +1,10 @@
+from django.contrib.auth.models import User
+
+from rest_framework import serializers
+
+
+# Serializers define the API representation.
+class UserSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = User
+        fields = ('url', 'username', 'email', 'is_staff')
