+++ apps/authentication/views.py
@@ -1,3 +1,11 @@
-from django.shortcuts import render
+from django.contrib.auth.models import User
 
-# Create your views here.
+from rest_framework import viewsets
+
+from apps.authentication.serializers import UserSerializer
+
+
+# ViewSets define the view behavior.
+class UserViewSet(viewsets.ModelViewSet):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
